examples/security/identity-sso/keycloak-python/requirements.txt
# Core dependencies
python-keycloak>=4.0.0
httpx>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# CLI
typer>=0.9.0
rich>=13.7.0

# FastAPI (optional - for API server example)
fastapi>=0.109.0
uvicorn>=0.27.0
python-jose[cryptography]>=3.3.0

# Token storage
keyring>=24.3.0
cryptography>=42.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
